integration_testing/requirements.txt
pytest>=7.2.0
pytest-xdist
pytest-cov
pytest-timeout
requests
djangorestframework==3.14.0
pytest-django==4.5.2
